requirements.txt
asgiref==3.8.1
Django==5.1
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.0
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
